chatbot/requirements.txt
# Web server
flask>=3.0.0

# RAG dependencies (same as parent project)
langchain==0.1.10
langchain-community>=0.0.25
langchain-openai==0.0.8
langchain-qdrant==0.1.3
qdrant-client>=1.10.1
openai>=1.3.0

# Environment variables
python-dotenv==1.0.0
